apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-marketplace-backend"
version = "0.1.0"
description = "FastAPI backend for agent marketplace MVP"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "passlib>=1.7.4,<2.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]
